@@ -3,7 +3,10 @@
 namespace ChrisKonnertz\DeepLy\HttpClient;
 
 /*
- * This exception might be thrown during an API call
+ * This exception might be thrown during an API call.
+ * The "message" and "code" property will be set.
+ *
+ * Learn more about the codes: http://deeply.test/demos/demo_translate.php
  */
 class CallException extends \RuntimeException
 {
