@@ -12,7 +12,8 @@ class DeepLyServiceProvider extends ServiceProvider
     {
         $this->app->bind('deeply', function()
         {
-            return new DeepLy();
+            // Create new DeepLy instance with API key from .env file
+            return new DeepLy(env('DEEPL_API_KEY'));
         });
     }
 
